requirements.txt
numpy
plotly
pyserial
